refactor test framework
